+++ python/GafferSceneTest/__init__.py
@@ -153,6 +153,7 @@ from .DeleteSetsTest import DeleteSetsTest
 from .UnencapsulateTest import UnencapsulateTest
 from .MotionPathTest import MotionPathTest
 from .FilterQueryTest import FilterQueryTest
+from .TransformQueryTest import TransformQueryTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *